refactor(UI): new tabs design, better product details display in add & edit mode & and modularize components
